Move mongo planner helpers into separate file
